+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Attendance entries can now be deleted individually from the status popup.
+
 ## [1.4.2] - 2026-08-27
 
 ### Fixed